refactor: remove explicit indices from kwk when they are monotonic
They were intended to make the mapping clearer, but in fact they hide
the result which can be seen more easily in VSCode when indices are not
there. Removing them makes no difference since they are assigned
monotoically by default.

Of course, we keep the indices when they are not monotinic.